Mise en place des boutons à la bonne position et configuration de ceux-ci pour effecter une tâche qui leur est assigné. Mise en place de fonction qui génère les traces directement à partir d'entiers (ControlTrain.c). Correction dans certaines librairies et commentaires.

+++ Programmes_LPC/StationDeGestion_TP_Train/src/police.h
@@ -11,16 +11,15 @@
 #include "LPC17xx.h"
 #include <cr_section_macros.h>
 #include "Ecran.h"
-#include <math.h>
 
 
 #define LETTER_WIDTH 8
 #define LETTER_HEIGHT 22
 
 void Write_char(char character, uint16_t x,uint16_t y, uint8_t *color);
-void Write_string(char * string,uint8_t x,uint8_t y, uint8_t *color_character);
-void Write_string_with_background(char * string,uint8_t x,uint8_t y, uint8_t *color_character,uint8_t *color_background);
-void Write_char_with_background(char character, uint8_t x,uint8_t y, uint8_t *color_letter, uint8_t *color_background);
-void Create_button(char * string,uint8_t x,uint8_t y,uint8_t hauteur,uint8_t largeur);
+void Write_string(char * string,uint16_t x,uint16_t y, uint8_t *color_character);
+void Write_string_with_background(char * string,uint16_t x,uint16_t y, uint8_t *color_character,uint8_t *color_background);
+void Write_char_with_background(char character, uint16_t x,uint16_t y, uint8_t *color_letter, uint8_t *color_background);
+void Create_button(char * string,uint16_t x,uint16_t y,uint8_t hauteur,uint8_t largeur);
 
 #endif /* POLICE_H_ */
